Remove fixture extractor
