+++ crates/karva_core/src/extensions/fixtures/mod.rs
@@ -3,7 +3,6 @@ use std::fmt::Display;
 use pyo3::{prelude::*, types::PyTuple};
 use ruff_python_ast::{Expr, StmtFunctionDef};
 
-pub mod extractor;
 pub mod finalizer;
 pub mod manager;
 
@@ -191,7 +190,7 @@ impl Fixture {
             .getattr(function_definition.name.to_string())
             .map_err(|e| e.to_string())?;
 
-        let try_karva = extractor::try_from_karva_function(
+        let try_karva = Self::try_from_karva_function(
             py,
             function_definition,
             &function,
@@ -205,7 +204,7 @@ impl Fixture {
             Err(e) => Some(e),
         };
 
-        let try_pytest = extractor::try_from_pytest_function(
+        let try_pytest = Self::try_from_pytest_function(
             py,
             function_definition,
             &function,
@@ -218,6 +217,109 @@ impl Fixture {
             Ok(None) => try_karva_err.map_or_else(|| Ok(None), Err),
             Err(e) => Err(e),
         }
+    }
+
+    pub(crate) fn try_from_pytest_function(
+        py: Python<'_>,
+        function_definition: &StmtFunctionDef,
+        function: &Bound<'_, PyAny>,
+        module_name: &str,
+        is_generator_function: bool,
+    ) -> Result<Option<Self>, String> {
+        let Some(found_name) =
+            get_attribute(function.clone(), &["_fixture_function_marker", "name"])
+        else {
+            return Ok(None);
+        };
+
+        let Some(scope) = get_attribute(function.clone(), &["_fixture_function_marker", "scope"])
+        else {
+            return Ok(None);
+        };
+
+        let Some(auto_use) =
+            get_attribute(function.clone(), &["_fixture_function_marker", "autouse"])
+        else {
+            return Ok(None);
+        };
+
+        let Some(function) = get_attribute(function.clone(), &["_fixture_function"]) else {
+            return Ok(None);
+        };
+
+        let name = if found_name.is_none() {
+            function_definition.name.to_string()
+        } else {
+            found_name.to_string()
+        };
+
+        let fixture_scope = fixture_scope(py, &scope, &name)?;
+
+        Ok(Some(Self::new(
+            QualifiedFunctionName::new(name, module_name.to_string()),
+            function_definition.clone(),
+            fixture_scope,
+            auto_use.extract::<bool>().unwrap_or(false),
+            function.into(),
+            is_generator_function,
+        )))
+    }
+
+    pub(crate) fn try_from_karva_function(
+        py: Python<'_>,
+        function_def: &StmtFunctionDef,
+        function: &Bound<'_, PyAny>,
+        module_name: &str,
+        is_generator_function: bool,
+    ) -> Result<Option<Self>, String> {
+        let Ok(py_function) = function
+            .clone()
+            .downcast_into::<python::FixtureFunctionDefinition>()
+        else {
+            return Ok(None);
+        };
+
+        let Ok(py_function_borrow) = py_function.try_borrow_mut() else {
+            return Ok(None);
+        };
+
+        let scope_obj = py_function_borrow.scope.clone();
+        let name = py_function_borrow.name.clone();
+        let auto_use = py_function_borrow.auto_use;
+
+        let fixture_scope = fixture_scope(py, scope_obj.bind(py), &name)?;
+
+        Ok(Some(Self::new(
+            QualifiedFunctionName::new(name, module_name.to_string()),
+            function_def.clone(),
+            fixture_scope,
+            auto_use,
+            py_function.into(),
+            is_generator_function,
+        )))
+    }
+}
+
+fn get_attribute<'a>(function: Bound<'a, PyAny>, attributes: &[&str]) -> Option<Bound<'a, PyAny>> {
+    let mut current = function;
+    for attribute in attributes {
+        let current_attr = current.getattr(attribute).ok()?;
+        current = current_attr;
+    }
+    Some(current.clone())
+}
+
+fn fixture_scope(
+    py: Python<'_>,
+    scope_obj: &Bound<'_, PyAny>,
+    name: &str,
+) -> Result<FixtureScope, String> {
+    if scope_obj.is_callable() {
+        resolve_dynamic_scope(py, scope_obj, name)
+    } else if let Ok(scope_str) = scope_obj.extract::<String>() {
+        FixtureScope::try_from(scope_str)
+    } else {
+        Err("Scope must be either a string or a callable".to_string())
     }
 }
